aktualizace cv05
oprava cv05 - každá trieda ma samostatný súbor
