+++ www/jezf00/cv05/database/database.php
@@ -30,80 +30,10 @@ abstract class Database implements DatabaseOperations
   }
 }
 
-class UsersDB extends Database
-{
-  public function create($args)
-  {
-    $this->save($args);
-    echo 'User ', $args['name'], ' age: ', $args['age'], ' was created', PHP_EOL;
-  }
+require './database/UsersDB.php'; 
+require './database/ProductsDB.php';
+require  './databese/OrdersDB.php';
 
-  public function fetch()
-  {
-    echo 'A user was fetched', PHP_EOL;
-  }
-
-  public function save($args)
-  {
-    file_put_contents($this->filePath(), $args['name'] . $this->getSeparator() . $args['age'] . PHP_EOL, FILE_APPEND);
-    echo 'A user was saved  ', PHP_EOL;
-  }
-
-  public function delete()
-  {
-    echo 'A user cannot be deleted', PHP_EOL;
-  }
-}
-
-class ProductsDB extends Database
-{
-  public function create($args)
-  {
-    $this->save($args);
-    echo 'Product ', $args['name'], ' $', $args['price'], ' was created', PHP_EOL;
-  }
-
-  public function fetch()
-  {
-    echo 'A product was fetched', PHP_EOL;
-  }
-
-  public function save($args)
-  {
-    file_put_contents($this->filePath(), $args['name'] . $this->getSeparator() . $args['price'] . PHP_EOL, FILE_APPEND);
-    echo 'A product was saved  ', PHP_EOL;
-  }
-
-  public function delete()
-  {
-    echo 'A product cannot be deleted', PHP_EOL;
-  }
-}
-
-class OrdersDB extends Database
-{
-  public function create($args)
-  {
-    echo 'Order no. ', $args['number'], ' was created', PHP_EOL;
-    $this->save($args);
-  }
-
-  public function fetch()
-  {
-    echo 'An order was fetched', PHP_EOL;
-  }
-
-  public function save($args)
-  {
-    file_put_contents($this->filePath(), $args['number'] . PHP_EOL, FILE_APPEND);
-    echo 'An order was saved  ', PHP_EOL;
-  }
-
-  public function delete()
-  {
-    echo 'An order cannot be deleted', PHP_EOL;
-  }
-}
 
 $users = new UsersDB();
 echo $users, PHP_EOL;
@@ -125,3 +55,4 @@ echo $orders, PHP_EOL;
 $orders->create(['number' => 15]);
 $orders->fetch();
 echo PHP_EOL;
+?>
